feat: add a true path sanitizing function to base World
BREAKING CHANGES:

The base world object no longer has a public `projectRoot` property.
Instead of constructing paths that way, developers should use the
`world.fullPath()` method to get the full path to a file or subfolder.
This should ensure that files outside the project root are not accessed.

Playwright and Browser world constructors no longer have the
sanitizeFilepath method. Instead, there is a new "sanitizePath"
method on the base World, which should be relatively safe even
for user input, avoiding path traversal and the like.